Move IFBoard Driver
Also some driveby fixes on system parameters
Import still needs pyserial and matplotlib for the server stuff which is unfortunate but can be corrected

Resolves #14
